refactor: remove LottoNumbers and replace it to Lotto
