+++ src/test/kotlin/lotto/WinningNumbersTest.kt
@@ -8,7 +8,7 @@ class WinningNumbersTest {
     @Test
     fun `verify `() {
         assertThrows<IllegalArgumentException> {
-            WinningNumbers(LottoNumbers(listOf(1, 2, 3, 4, 5, 6)), 1)
+            WinningNumbers(Lotto(listOf(1, 2, 3, 4, 5, 6)), 1)
         }
     }
 }
